fix(entities): add data message and message entities, modify notification entity
